refactor: WIP #29 routing for basic components
