+++ .babelrc
@@ -1,3 +1,3 @@
 {
-"presets": ["react-native"]
+"presets": ["react-native", "react-native-stage-0/decorator-support"]
 }
