+++ package.json
@@ -9,7 +9,9 @@
     "test:watch": "nodemon -q -x 'npm test'"
   },
   "dependencies": {
+    "@exponent/ex-navigation": "^2.1.5",
     "axios": "^0.15.2",
+    "babel-preset-react-native-stage-0": "^1.0.1",
     "lodash": "^4.17.2",
     "moment": "^2.17.0",
     "react": "15.4.1",
