+++ src/js/routes.js
@@ -1,0 +1,14 @@
+import { createRouter } from '@exponent/ex-navigation';
+import AlbumsContainer from './containers/albums';
+import CalendarContainer from './containers/calendar';
+import FeedContainer from './containers/feed';
+import ProfileContainer from './containers/profile';
+
+const Router = createRouter(() => ({
+  albums: () => AlbumsContainer,
+  calendar: () => CalendarContainer,
+  feed: () => FeedContainer,
+  profile: () => ProfileContainer
+}));
+
+export default Router;
